fix: Ensures the directory is deleted only if the option is set
From e2e test it have a check for this and it throws me error so i decided to add this to prevent cases when option is undefined

+++ src/visitors/renderVisitor.ts
@@ -7,7 +7,7 @@ import { getRenderMapVisitor } from './getRenderMapVisitor';
 
 export function renderVisitor(path: string, options: RenderOptions) {
     return rootNodeVisitor(root => {
-        if (options.deleteFolderBeforeRendering ?? true) {
+        if (options && (options.deleteFolderBeforeRendering ?? true)) {
             deleteDirectory(path);
         }
